rename lspconfig
