+++ nvim/init.lua
@@ -1,9 +1,9 @@
 require('base')
 require('keymap')
 require('autocmd')
+require('lspconfig')
 require('plugin.gitsigns')
 require('plugin.fzf-lua')
-require('plugin.lspconfig')
 require('plugin.blink')
 require('plugin.bufferline')
 require('colorscheme')
